refactor[#123]: 수정사항 반영
